pyproject.toml
[tool.poetry]
name = "ltgenius-api"
version = "0.1.0"
description = "Lockthreat APIs"
authors = ["Cirruslabs <email@cirruslabs.io>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "^2.7.4"
openai = "^1.42.0"
arrow = "^1.3.0"
pydantic-settings = "^2.4.0"
python-decouple = "^3.8"
azure-identity = "^1.18.0"
azure-keyvault-secrets = "^4.8.0"
fastapi = "^0.115.0"
python-multipart = "0.0.10"
mysql-connector-python = "^9.0.0"
scikit-learn = "1.5.2"
langchain = "^0.3.2"
langchain-openai = "^0.2.1"
langchain-community = "^0.3.1"
langchain-experimental = "^0.3.2"
azure-storage-blob = "^12.23.1"
langchain-ollama = "^0.2.3"
streamlit = "^1.41.1"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
mypy = "^1.4.1"
ruff = "^0.0.278"
black = "^23.7.0"
syrupy = "^4.0.2"
notebook = "^7.2.1"
pandas = "^2.2.2"
nbdev = "^2.3.34"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
  "E",  # pycodestyle
  "F",  # pyflakes
  "I",  # isort
]

[tool.mypy]
ignore_missing_imports = "True"
disallow_untyped_defs = "True"


